Changes to DMC and MC algorithm to make them faster
The step function of the MC algorithm was reworked, calculating the
necessary normals therefore removing the need to buffer all vertices to
calculate them.
There are errors in the normal calculation for the DMC algorithm now,
which have to be investigated.
Also both codes will have to be cleaned up.